Update function library, introduce trampoline
* add LICENSE header, making this the first version people are allowed
  to actually use. yay.
* remove reliance on global context variables across the board
* shrink function names to make space for arguments
* remove wrappers that call functions with specific global variables
  as argument
* combine into generic functions where possible

The big functional update is the logic inside the pam_pf anchor. The
main anchor now only contains:
* a table with the source addresses registered for the user
* a conditional sub-anchor for the regular ruleset that is the same
  for every user
* a conditional sub-anchor for the extended per-user ruleset

With only the conditional sub-anchors, an efficient early-exit option
is given to pf that reduces the ruleset evaluations.

Also introduce %LABEL% inside the loaded rulesets, which will be
replaced by pam_pf with the actual label prior to loading. States
opened by rules with this label will be killed at cleanup.